api/urls.py
from django.urls import path
from api import views
from rest_framework_simplejwt.views import (TokenObtainPairView, TokenRefreshView)

urlpatterns = [
    path('users/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('users/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('', views.get_response),
    path('records/', views.get_records),
    path('records/<uuid:pk>/', views.get_record)]
